- Declared asn1 inlined conversions functions as static to avoid errors at link time - Forward initial context setup req message to RRC - Updated some data types related to initial context setup request
git-svn-id: http://svn.eurecom.fr/openair4G/trunk@4454 818b1a75-f10b-46b9-bf7c-635c3b92a50f

+++ openair-cn/COMMON/common_types.h
@@ -74,7 +74,7 @@ typedef enum {
     NAM_MAX,
 } network_access_mode_t;
 
-typedef uint32_t bitrate_t;
+typedef uint64_t bitrate_t;
 
 typedef uint8_t  ebi_t;
 typedef char*    APN_t;
